Move WebSocketEvent to entity/streaming package
