+++ bigbone-rx/src/main/kotlin/social/bigbone/rx/RxStreamingMethods.kt
@@ -3,14 +3,14 @@ package social.bigbone.rx
 import io.reactivex.rxjava3.core.BackpressureStrategy
 import io.reactivex.rxjava3.core.Flowable
 import social.bigbone.MastodonClient
-import social.bigbone.api.Closed
-import social.bigbone.api.Closing
-import social.bigbone.api.Failure
-import social.bigbone.api.GenericMessage
-import social.bigbone.api.Open
-import social.bigbone.api.StreamEvent
 import social.bigbone.api.WebSocketCallback
-import social.bigbone.api.WebSocketEvent
+import social.bigbone.api.entity.streaming.MastodonApiEvent.GenericMessage
+import social.bigbone.api.entity.streaming.MastodonApiEvent.StreamEvent
+import social.bigbone.api.entity.streaming.TechnicalEvent.Closed
+import social.bigbone.api.entity.streaming.TechnicalEvent.Closing
+import social.bigbone.api.entity.streaming.TechnicalEvent.Failure
+import social.bigbone.api.entity.streaming.TechnicalEvent.Open
+import social.bigbone.api.entity.streaming.WebSocketEvent
 import social.bigbone.api.method.StreamingMethods
 import java.io.Closeable
