+++ bigbone/src/main/kotlin/social/bigbone/MastodonClient.kt
@@ -11,17 +11,17 @@ import okhttp3.Response
 import okhttp3.WebSocket
 import okhttp3.WebSocketListener
 import okio.ByteString
-import social.bigbone.api.Closed
-import social.bigbone.api.Closing
-import social.bigbone.api.Failure
-import social.bigbone.api.GenericMessage
-import social.bigbone.api.Open
 import social.bigbone.api.Pageable
-import social.bigbone.api.StreamEvent
 import social.bigbone.api.WebSocketCallback
 import social.bigbone.api.entity.data.InstanceVersion
+import social.bigbone.api.entity.streaming.MastodonApiEvent.GenericMessage
+import social.bigbone.api.entity.streaming.MastodonApiEvent.StreamEvent
 import social.bigbone.api.entity.streaming.ParsedStreamEvent.Companion.toStreamEvent
 import social.bigbone.api.entity.streaming.RawStreamEvent
+import social.bigbone.api.entity.streaming.TechnicalEvent.Closed
+import social.bigbone.api.entity.streaming.TechnicalEvent.Closing
+import social.bigbone.api.entity.streaming.TechnicalEvent.Failure
+import social.bigbone.api.entity.streaming.TechnicalEvent.Open
 import social.bigbone.api.exception.BigBoneClientInstantiationException
 import social.bigbone.api.exception.BigBoneRequestException
 import social.bigbone.api.exception.InstanceVersionRetrievalException
